pyproject.toml
[project]
name = "yara-malware-scanner"
version = "1.0.0"
description = "Hệ thống phát hiện mã độc sử dụng YARA rules và hash-based detection"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.0.0",
    "yara-python>=4.5.0",
    "py7zr>=0.20.0",
    "rarfile>=4.0",
    "olefile>=0.47",
]

[project.urls]
Homepage = "https://github.com/2312702-NBTKNguyen/LapTrinhPython_DoAnGiuaKy_Yara"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["malware_scanner"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[tool.uv.pip]
index-strategy = "unsafe-best-match"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
